Projects and Voly pages and Bug Fixes
Added projects page, which will be a home for any big projects that have their own web page dedicated to them. The first of which is Voly, which has a web page currently. It's almost done, but I wanted to get this out there for potential jobs I will be applying to soon.

Fixed a bug where clicking the navButtons to navigate through pictures wasn't working.